Screenshot capturing program
When im watching video lectures(Online) that time I want to take screenshot of current window, so I have to 2 option one is using PtrSc (print screen Button of keyboard and then paste it into paint) and another one is snipping tool , both has disadvantage of time and stop work ... As I used Ubuntu(long long ago) In that there is such facility to capture screen directly and it save on at specified location. So here I created such python program which capture current screen and save to specified folder with current time as a name.